refractor

+++ packages/plugins/data-collect/image-classification-data-collect/src/index.ts
@@ -41,6 +41,7 @@ const imageClassDataCollect: DataCollectType = async (args: ArgsType): Promise<v
 
   assert.ok(extention[extention.length - 1] === 'zip', 'The dataset provided should be a zip file');
 
+  let isDownload = false;
   if (/^file:\/\/.*/.test(url)) {
     url = url.substring(7);
   } else {
@@ -48,6 +49,7 @@ const imageClassDataCollect: DataCollectType = async (args: ArgsType): Promise<v
     console.log('downloading dataset ...')
     await download(url, targetPath);
     url = targetPath;
+    isDownload = true;
   }
 
   const imageDir = path.join(dataDir, 'images');
@@ -65,7 +67,9 @@ const imageClassDataCollect: DataCollectType = async (args: ArgsType): Promise<v
     fs.moveSync(imagePath, path.join(annotationDir, imageName), { overwrite: true });
   });
 
-  fs.removeSync(url);
+  if (isDownload) {
+    fs.removeSync(url);
+  }
   fs.removeSync(path.join(dataDir, 'images'));
 }
